Separo test cases de query en módulos aparte
    refs:#91
