+++ series_tiempo_ar_api/apps/api/tests/es_query_tests.py
@@ -102,106 +102,3 @@ class QueryTest(TestCase):
     @raises(QueryError)
     def test_try_collapse(self):
         self.query.add_collapse(interval='year', agg='avg')
-
-
-class CollapseQueryTests(TestCase):
-
-    start = 10
-    limit = 15
-
-    single_series = SERIES_NAME
-    rep_mode = 'value'
-    series_periodicity = 'month'
-
-    @classmethod
-    def setUpClass(cls):
-        setup_database()
-        super(CollapseQueryTests, cls).setUpClass()
-
-    def setUp(self):
-        self.query = CollapseQuery(index=settings.TEST_INDEX)
-
-    @raises(QueryError)
-    def test_execute_empty(self):
-        data = self.query.run()
-
-        self.assertFalse(data)
-
-    def test_execute_single(self):
-        self.query.add_series(self.single_series, self.rep_mode, self.series_periodicity)
-
-        data = self.query.run()
-        self.assertTrue(data)
-
-    def test_start_limit(self):
-        self.query.add_series(self.single_series, self.rep_mode, self.series_periodicity)
-        self.query.add_pagination(self.start, self.limit)
-        data = self.query.run()
-
-        self.assertEqual(len(data), self.limit)
-
-    def test_init_from_other(self):
-        other_query = ESQuery(index=settings.TEST_INDEX)
-        other_query.add_series(self.single_series, self.rep_mode, self.series_periodicity)
-        self.query = CollapseQuery(settings.TEST_INDEX, other=other_query)
-        self.query.add_collapse()
-        data = self.query.run()
-        self.assertTrue(data)
-
-    def test_add_collapse(self):
-        """Testea que luego de agregar un collapse default, los
-        resultados sean anuales, es decir cada uno a un año de
-        diferencia con su anterior"""
-        self.query.add_series(self.single_series, self.rep_mode, self.series_periodicity)
-        self.query.add_collapse()
-        data = self.query.run()
-        prev_timestamp = None
-        for row in data:
-            timestamp = row[0]
-            parsed_timestamp = iso8601.parse_date(timestamp)
-            if not prev_timestamp:
-                prev_timestamp = parsed_timestamp
-                continue
-            delta = relativedelta(parsed_timestamp, prev_timestamp)
-            self.assertTrue(delta.years == 1, timestamp)
-            prev_timestamp = parsed_timestamp
-
-    def test_collapse_custom_params(self):
-        self.query.add_series(self.single_series, self.rep_mode, self.series_periodicity)
-        self.query.add_collapse(interval='quarter')
-        data = self.query.run()
-        prev_timestamp = None
-        for row in data:
-            timestamp = row[0]
-            parsed_timestamp = iso8601.parse_date(timestamp)
-            if not prev_timestamp:
-                prev_timestamp = parsed_timestamp
-                continue
-            delta = relativedelta(parsed_timestamp, prev_timestamp)
-            self.assertTrue(delta.months == 3, timestamp)
-            prev_timestamp = parsed_timestamp
-
-    def test_add_two_collapses(self):
-        """Esperado: El segundo collapse overridea el primero"""
-        self.query.add_series(self.single_series, self.rep_mode, self.series_periodicity)
-        self.query.add_collapse(interval='quarter')
-        self.query.add_collapse(interval='year')
-        data = self.query.run()
-
-        prev_timestamp = None
-        for row in data:
-            timestamp = row[0]
-            parsed_timestamp = iso8601.parse_date(timestamp)
-            if not prev_timestamp:
-                prev_timestamp = parsed_timestamp
-                continue
-            delta = relativedelta(parsed_timestamp, prev_timestamp)
-            self.assertTrue(delta.years == 1, timestamp)
-            prev_timestamp = parsed_timestamp
-
-    def test_init_from_other_collapse_query(self):
-        other_query = CollapseQuery(index=settings.TEST_INDEX)
-        other_query.add_series(self.single_series, self.rep_mode, self.series_periodicity)
-        self.query = CollapseQuery(index=settings.TEST_INDEX, other=other_query)
-        data = self.query.run()
-        self.assertTrue(data)
